Dia 15/12 - Versão 1 finalizada de GamesGate

+++ App/Models/BibliotecaModel.cs
@@ -1,9 +1,4 @@
 ﻿using Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace App.Models
 {
@@ -25,7 +20,7 @@ namespace App.Models
         public int Id { get; set; }
         public int IdJogo { get; set; }
         public string? Jogo { get; set; }
-        public double HorasJogadas { get; set; }
+        public float HorasJogadas { get; set; }
         public int IdBiblioteca { get; set; }
     }
 }
